rename Comparison folder
