+++ src/pages/Сomparison/ComparisonPage.module.scss
