+++ src/pages/Сomparison/index.ts
