+++ src/pages/Сomparison/СomparisonPage.tsx
